vm code refactoring
