+++ src/backend/VirtualMachine/build.rs
@@ -6,7 +6,8 @@ use std::path::PathBuf;
 fn main() {
   // Tell cargo to tell rustc to link the system bzip2
   // shared library.
-  println!("cargo:rustc-link-lib=DbEngineInterface");
+  // FIXME: uncomment this
+  // println!("cargo:rustc-link-lib=DbEngineInterface");
   println!("cargo:rustc-link-search=../lib");
 
   // Tell cargo to invalidate the built crate whenever the wrapper changes
