+++ src/backend/VirtualMachine/src/lib.rs
@@ -4,7 +4,6 @@
 #![allow(non_camel_case_types)]
 pub mod VirtualMachine;
 pub mod wrapper;
-use wrapper::wrapper::cLayer;
 
 extern crate num;
 #[macro_use]
